HomuraServerProtocolをclasses/procotol.pyに移動
長かったので一応分けました
