+++ classes/__init__.py
@@ -1,2 +1,3 @@
 from .booltool import toBool
 from .lognk import log
+from .protocol import HomuraServerProtocol
